honksquad: Split reagent-analysis engine out of HealthAnalyzerReagentSystem (#858)
Behavior-preserving extraction of two responsibilities tangled into the
429-line scanner file:

- ReagentDoseThresholdAnalyzer: the reagent-analysis engine
  (GetDoseThresholds + ClassifyEffect/ClassifyDamageValues/SelfBounds and
  its _thresholdCache, plus the prototype-reload cache invalidation). Now
  its own EntitySystem so the classifier is reusable by other tooling
  (chem scanners, pharmacist tools) and testable in isolation.

- SolutionAggregator: the AddSolution / AddOrganSolutions family that
  walks bloodstream / metabolites / stomach / lung solutions into the
  reagent group list. BuildState is now a thin wrapper that delegates
  group construction and attaches the display name.

HealthAnalyzerReagentSystem drops from 429 to ~170 lines (scanner
lifecycle + UI-state plumbing only). Threshold tests now resolve
ReagentDoseThresholdAnalyzer directly.
